+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warnings about locks are now reliably printed every 30 seconds
 - We now make sure Beaker jobs have the latest version of beaker-py, so that we're compatible with the latest API changes.
 - Stopping early now works when the metric doesn't change at all.
+- Fixed bug with `FromParams` which didn't handle variable length tuples correctly.
 
 ### Changed